+++ Sources/VCore/Services and Managers/Presentation Host/Geometry/PresentationHostGeometryReaderSafeAreaInsetsEnvironmentValue.swift
@@ -7,20 +7,6 @@
 
 import SwiftUI
 
-// MARK: - Presentation Host Geometry Reader Safe Area Insets Extension
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-@available(visionOS, unavailable)
-extension View {
-    func presentationHostGeometryReaderSafeAreaInsets(
-        _ presentationHostGeometryReaderSafeAreaInsets: EdgeInsets
-    ) -> some View {
-        self
-            .environment(\.presentationHostGeometryReaderSafeAreaInsets, presentationHostGeometryReaderSafeAreaInsets)
-    }
-}
-
 // MARK: - Presentation Host Geometry Reader Safe Area Insets Environment Value
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
@@ -28,17 +14,5 @@ extension View {
 @available(visionOS, unavailable)
 extension EnvironmentValues {
     /// Presentation Host's `GeometryReader` safe area insets associated with the environment.
-    public var presentationHostGeometryReaderSafeAreaInsets: EdgeInsets {
-        get { self[PresentationHostGeometryReaderSafeAreaInsetsEnvironmentKey.self] }
-        set { self[PresentationHostGeometryReaderSafeAreaInsetsEnvironmentKey.self] = newValue }
-    }
-}
-
-// MARK: - Presentation Host Geometry Reader Safe Area Insets Environment Key
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-@available(visionOS, unavailable)
-struct PresentationHostGeometryReaderSafeAreaInsetsEnvironmentKey: EnvironmentKey {
-    static let defaultValue: EdgeInsets = .init()
+    @EnvironmentValueGeneration public var presentationHostGeometryReaderSafeAreaInsets: EdgeInsets = .init()
 }