+++ Sources/VCore/Services and Managers/Presentation Host/Presentation Mode/PresentationHostPresentationModeEnvironmentValue.swift
@@ -7,20 +7,6 @@
 
 import SwiftUI
 
-// MARK: - Presentation Host Presentation Mode Extension
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-@available(visionOS, unavailable)
-extension View {
-    func presentationHostPresentationMode(
-        _ presentationHostPresentationMode: PresentationHostPresentationMode?
-    ) -> some View {
-        self
-            .environment(\.presentationHostPresentationMode, presentationHostPresentationMode)
-    }
-}
-
 // MARK: - Presentation Host Presentation Mode Environment Value
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
@@ -28,17 +14,5 @@ extension View {
 @available(visionOS, unavailable)
 extension EnvironmentValues {
     /// Presentation Host's presentation mode of the `View` associated with the environment.
-    public var presentationHostPresentationMode: PresentationHostPresentationMode? {
-        get { self[PresentationHostPresentationModeEnvironmentKey.self] }
-        set { self[PresentationHostPresentationModeEnvironmentKey.self] = newValue }
-    }
-}
-
-// MARK: - Presentation Host Presentation Mode Environment Key
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-@available(visionOS, unavailable)
-private struct PresentationHostPresentationModeEnvironmentKey: EnvironmentKey {
-    static let defaultValue: PresentationHostPresentationMode? = nil
+    @EnvironmentValueGeneration public var presentationHostPresentationMode: PresentationHostPresentationMode?
 }