+++ Sources/VCore/Views/SwiftUI/View Resetting Container (Observable Object)/ViewResetterOOEnvironmentValue.swift
@@ -7,28 +7,10 @@
 
 import SwiftUI
 
-// MARK: - View Resetter Extension (Observable Object)
-extension View {
-    func viewResetterOO(
-        _ viewResetter: ViewResetterOO
-    ) -> some View {
-        self
-            .environment(\.viewResetterOO, viewResetter)
-    }
-}
-
 // MARK: - View Resetter Environment Value (Observable Object)
 extension EnvironmentValues {
     /// `ViewResetterOO` of the `View` associated with the environment.
     ///
     /// Since `ViewResetterOO` is a reference type, `View` updates won't be triggered.
-    public var viewResetterOO: ViewResetterOO? {
-        get { self[ViewResetterEnvironmentKeyOO.self] }
-        set { self[ViewResetterEnvironmentKeyOO.self] = newValue }
-    }
-}
-
-// MARK: - View Resetter Environment Key (Observable Object)
-private struct ViewResetterEnvironmentKeyOO: EnvironmentKey {
-    static let defaultValue: ViewResetterOO? = nil
+    @EnvironmentValueGeneration public var viewResetterOO: ViewResetterOO?
 }